Also regroup targets by folder
